Add dev-mode frontend option for debug purpose and fix magnitude computation
A new option has been added to the frontend which enables the system to hand over models, which contain errors, to the backend.
While this may not be designed for the end-used, it can be used during the debuging of the software to find errors.
A bug has been fixed which resulted in exceptions whenever the GSL reference converter tried to get a astropy unit from the
set of defined units. This should no longer happen.